Got Sphinx documentation working.
The API documentation can now be generated by Sphinx. The html template
uses Bootstrap, to keep the look and feel consistent with the application.
The settings are slightly different, but that just requires some tweaking
of parameters.

The Sphinx setup still needs some refinement, but at least we don't
have to dig through code to find out how to use stuff.